Fix: asset > git > submodule > local.sh | [ -e ] --> [[ -e ]]

+++ asset/git/submodule/local.sh
@@ -7,9 +7,9 @@ HOME=$(realpath ~/)
 
 # Replace
 sed -i -e "s|https://github.com/onepiece-framework/|${HOME}/repo/op/|g" .gitmodules
-sed -i -e "s|/unit-|/unit/|g"          .gitmodules
-sed -i -e "s|/module-|/module/|g"      .gitmodules
-sed -i -e "s|/layout-|/layout/|g"      .gitmodules
+sed -i -e "s|/unit-|/unit/|g"          .gitmodules # old path, Retained for compatibility.
+sed -i -e "s|/module-|/module/|g"      .gitmodules # old path, Retained for compatibility.
+sed -i -e "s|/layout-|/layout/|g"      .gitmodules # old path, Retained for compatibility.
 sed -i -e "s|/op-core.git|/core.git|g" .gitmodules
 sed -i -e "s|/op-unit-|/unit/|g"       .gitmodules
 sed -i -e "s|/op-module-|/module/|g"   .gitmodules
@@ -20,4 +20,6 @@ sed -i -e "s|/op-webpack-|/webpack/|g" .gitmodules
 git submodule sync
 
 # Delete garbage files.
-rm .gitmodules-e
+if [[ -e '.gitmodules-e' ]];then
+  rm .gitmodules-e
+fi
